pull saveWithStatusChange to own file
